refactor(problem 36): modularize for all problem
how
- extract solution, test cases
